requirements.txt
streamlit~=1.45.0
langchain==0.3.25
langchain-openai~=0.3.16
langchain-community~=0.3.23
duckduckgo-search==3.9.6
lxml-html-clean>=0.4.2
beautifulsoup4~=4.13.4
python-dotenv~=1.1.0
pytest==8.0.2
requests~=2.32.3
langchain-core~=0.3.59
bs4~=0.0.2
newspaper3k~=0.2.8
